Fix f2fs<->ext4 switch on /data partition

+++ partition.cpp
@@ -1160,7 +1160,7 @@ bool TWPartition::UnMount(bool Display_Error) {
 	}
 }
 
-bool TWPartition::Wipe(string New_File_System) {
+bool TWPartition::Wipe(string New_File_System, bool Force_Format) {
 	bool wiped = false, update_crypt = false;
 	int check;
 	string Layout_Filename = Mount_Point + "/.layout_version";
@@ -1185,13 +1185,13 @@ bool TWPartition::Wipe(string New_File_System) {
 	else
 		unlink("/.layout_version");
 
-	if (Has_Data_Media) {
+	if (Has_Data_Media && !Force_Format) {
 		wiped = Wipe_Data_Without_Wiping_Media();
 	} else {
 
 		DataManager::GetValue(TW_RM_RF_VAR, check);
 
-		if (check || Use_Rm_Rf)
+		if ((!check || Use_Rm_Rf) && !Force_Format)
 			wiped = Wipe_RMRF();
 		else if (New_File_System == "ext4")
 			wiped = Wipe_EXT4();
